login admin, confirmacion de cambios

+++ application/controllers/Administrador.php
@@ -14,23 +14,73 @@ class Administrador extends CI_Controller{
 		$this->load->view("administrador/index");
 		
 	}
+	public function getAdmin(){
+		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
+		session_name ( "cineProyecto" );
+		ini_set ( "session.cookie_lifetime", "7200" );
+		ini_set ( "session.gc_maxlifetime", "7200" );
+		session_start ();
+		
+		$user = $_GET['user'];
+		$pass = $_GET['pass'];
+		
+		$fila = R::getAll("SELECT * FROM usuario WHERE admin='SI' and user = :correo", array(':correo'=>$user));
+		if(count($fila)!="0"){
+			
+			$hash=$fila['0']['hash'];
+			if(password_verify($pass, $hash)){
+				
+				$_SESSION['idUserAdmin']['id']=$fila['0']['id'];
+				$_SESSION['idUserAdmin']['nombre']=$fila['0']['nombre'];
+				echo "OK";
+			}else{
+				echo "FAIL";
+			}
+		}else{
+			echo "FAIL";
+		}
+
+
+		R::close();
+		
+		
+	}
+
+	public function desconectar(){
+		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
+		session_name ( "cineProyecto" );
+		ini_set ( "session.cookie_lifetime", "7200" );
+		ini_set ( "session.gc_maxlifetime", "7200" );
+		session_start ();
+		
+		unset($_SESSION["idUserAdmin"]);
+		$this->load->view("administrador/index");
+		R::close();
+	}
+
 	public function principal(){
 		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
 		session_name ( "cineProyecto" );
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-		$this->load->model('usuarios_model');
-		$datos['cUsuarios']= $this->usuarios_model->getCantidadUsuarios();
-		$this->load->model('factura_model');
-		$datos['ingresosTotales']= $this->factura_model->ingresosTotales();
-		$this->load->model('entradareserva_model');
-		$datos['datoCant']=$this->entradareserva_model->datosCantidadSemana();
-		$this->load->model('pelicula_model');
-		$datos['topPelis']=$this->pelicula_model->getPeliculasTop();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model('usuarios_model');
+			$datos['cUsuarios']= $this->usuarios_model->getCantidadUsuarios();
+			$this->load->model('factura_model');
+			$datos['ingresosTotales']= $this->factura_model->ingresosTotales();
+			$this->load->model('entradareserva_model');
+			$datos['datoCant']=$this->entradareserva_model->datosCantidadSemana();
+			$this->load->model('pelicula_model');
+			$datos['topPelis']=$this->pelicula_model->getPeliculasTop();
+			$this->template->load("plantillaAdmin","administrador/principal", $datos);
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/principal", $datos);
+		
 		//$this->template->load("plantilla","administrador/principal");
 		// $this->load->view("prueba/Prue",$datos);
 		//$this->load->view("entrada/aConfirmar");
@@ -43,14 +93,21 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
 
-		$this->load->model("cartelera_model");
-		$datos['peliculas'] = $this->cartelera_model->getCartelera();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("cartelera_model");
+			$datos['peliculas'] = $this->cartelera_model->getCartelera();
 
-		$this->load->model("sala_model");
-		$datos['salas'] = $this->sala_model->getSalas();
+			$this->load->model("sala_model");
+			$datos['salas'] = $this->sala_model->getSalas();
+			$this->template->load("plantillaAdmin","administrador/crearSesiones",$datos);
+		}else{
+			$this->load->view("administrador/error");
+		}
+
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/crearSesiones",$datos);
+		
 		//$this->load->view("entrada/aConfirmar");
 	}
 	public function crearSesionesPost(){
@@ -59,13 +116,16 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("sesion_model");
+			$this->sesion_model->crearSesiones();
+			$this->template->load("plantillaAdmin","administrador/crearSesionesPost");
+		}else{
+			$this->load->view("administrador/error");
+		}
 
-		$this->load->model("sesion_model");
-		$this->sesion_model->crearSesiones();
-		
 		R::close();
 
-		$this->template->load("plantillaAdmin","administrador/crearSesionesPost");
 		//$this->load->view("entrada/aConfirmar");
 	}
 
@@ -75,12 +135,16 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->template->load("plantillaAdmin","administrador/crearPelicula");
+		}else{
+			$this->load->view("administrador/error");
+		}
 		
 		
 		R::close();
 		
-		$this->template->load("plantillaAdmin","administrador/crearPelicula");
+		
 		
 		//$this->template->load("plantilla","administrador/crearPelicula");
 
@@ -92,15 +156,19 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-
-		$this->load->model("pelicula_model");
-		$this->load->model("cartelera_model");
-		$data['idPeli']=$this->pelicula_model->crearPelicula();
-		print_r($data["idPeli"]);
-		if(isset($_POST["cartelera"])){
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$this->load->model("cartelera_model");
+			$data['idPeli']=$this->pelicula_model->crearPelicula();
+			print_r($data["idPeli"]);
+			if(isset($_POST["cartelera"])){
 			
-			$this->cartelera_model->anadirId($data['idPeli']);
+				$this->cartelera_model->anadirId($data['idPeli']);
+			}
+		}else{
+			$this->load->view("administrador/error");
 		}
+		
 		R::close();
 		$this->template->load("plantillaAdmin","administrador/crearPeliculaPost");
 	}
@@ -111,16 +179,21 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$data['peliculas']=$this->pelicula_model->getTodasPeliculas();
 
+			$this->load->model("cartelera_model");
+			$data['cartelera']=$this->cartelera_model->getCarteleraId();
+			$this->template->load("plantillaAdmin","administrador/cartelera",$data);
+		}else{
+			$this->load->view("administrador/error");
+		}
 
-		$this->load->model("pelicula_model");
-		$data['peliculas']=$this->pelicula_model->getTodasPeliculas();
-
-		$this->load->model("cartelera_model");
-		$data['cartelera']=$this->cartelera_model->getCarteleraId();
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/cartelera",$data);
+		
 	}
 
 	public function carteleraEliminar($id){
@@ -129,12 +202,17 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-
-		$this->load->model("cartelera_model");
-		$this->cartelera_model->eliminarId($id);
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("cartelera_model");
+			$this->cartelera_model->eliminarId($id);
+			$this->template->load("plantillaAdmin","administrador/carteleraEliminar");
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/carteleraEliminar");
+		
 	}
 
 	public function carteleraAnadir($id){
@@ -143,12 +221,16 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-
-		$this->load->model("cartelera_model");
-		$this->cartelera_model->anadirId($id);
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("cartelera_model");
+			$this->cartelera_model->anadirId($id);
+			$this->template->load("plantillaAdmin","administrador/carteleraAnadir");
+		}else{
+			$this->load->view("administrador/error");
+		}
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/carteleraAnadir");
+		
 	}
 
 	//SELECT SUM(er.precio) Precio, f.id, u.nombre, COUNT(er.asientos_id) Asientos FROM entradareserva er, entradareserva_factura erf, factura f, usuario u where erf.entradareserva_id = er.id AND f.id = erf.factura_id AND u.id = f.usuario_id GROUP BY erf.factura_id
@@ -161,12 +243,17 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-
-		$this->load->model("factura_model");
-		$data['informe'] = $this->factura_model->informeFactura();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("factura_model");
+			$data['informe'] = $this->factura_model->informeFactura();
+			$this->template->load("plantillaAdmin","administrador/informeFactura",$data);
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/informeFactura",$data);
+		
 	}
 
 	public function pruebaSesiones(){
@@ -175,11 +262,17 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-		$this->load->model("sesion_model");
-		$data['sesi']=$this->sesion_model->pruebaSesionSelect();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("sesion_model");
+			$data['sesi']=$this->sesion_model->pruebaSesionSelect();
+			$this->template->load("plantillaAdmin","administrador/prueba",$data);
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/prueba",$data);
+		
 	}
 
 	public function modificarPelicula(){
@@ -188,11 +281,17 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-		$this->load->model("pelicula_model");
-		$data['peliculas']=$this->pelicula_model->getTodasPeliculas();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$data['peliculas']=$this->pelicula_model->getTodasPeliculas();
+			$this->template->load("plantillaAdmin","administrador/modificarPelicula",$data);
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/modificarPelicula",$data);
+		
 		//$this->load->view("administrador/modificarPelicula");
 	}
 
@@ -202,23 +301,78 @@ class Administrador extends CI_Controller{
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
-		$this->load->model("pelicula_model");
-		$data['pelicula']=$this->pelicula_model->getPeliculaPorId($id);
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$data['pelicula']=$this->pelicula_model->getPeliculaPorId($id);
+			$data['id']=$id;
+			$this->template->load("plantillaAdmin","administrador/modificarPeliculaGet",$data);
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/modificarPeliculaGet",$data);
+		
 	}
 
-	public function modificar(){
+	public function modificarPeliculaPost($id){
 		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
 		session_name ( "cineProyecto" );
 		ini_set ( "session.cookie_lifetime", "7200" );
 		ini_set ( "session.gc_maxlifetime", "7200" );
 		session_start ();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$this->pelicula_model->modificarPelicula($id);
+			$this->template->load("plantillaAdmin","administrador/modificarPeliculaPost");
+		}else{
+			$this->load->view("administrador/error");
+		}
 		
 		
 		R::close();
-		$this->template->load("plantillaAdmin","administrador/a");
+		
+	}
+
+	public function eliminarPelicula(){
+		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
+		session_name ( "cineProyecto" );
+		ini_set ( "session.cookie_lifetime", "7200" );
+		ini_set ( "session.gc_maxlifetime", "7200" );
+		session_start ();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$data['peliculas']=$this->pelicula_model->getTodasPeliculas();
+			$this->template->load("plantillaAdmin","administrador/eliminarPelicula",$data);
+		
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
+		R::close();
+		
+		
+		
+	}
+
+	public function eliminarPeliculaGet($id){
+		R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
+		session_name ( "cineProyecto" );
+		ini_set ( "session.cookie_lifetime", "7200" );
+		ini_set ( "session.gc_maxlifetime", "7200" );
+		session_start ();
+		if(isset($_SESSION["idUserAdmin"])){
+			$this->load->model("pelicula_model");
+			$this->pelicula_model->eliminarPelicula($id);
+			$this->template->load("plantillaAdmin","administrador/eliminarPeliculaGet");
+		}else{
+			$this->load->view("administrador/error");
+		}
+		
+		R::close();
+		
+		
+		
 	}
 
 
@@ -241,6 +395,25 @@ class Administrador extends CI_Controller{
 //	 	// $this->load->view("prueba/Prue",$datos);
 //	 	//$this->load->view("entrada/aConfirmar");
 //	 }
+
+		//  public function crearAdmin(){
+	 // 	R::setup('mysql:host=localhost;dbname=proyecto', 'root', '');
+	 // 	session_name ( "cineProyecto" );
+	 // 	ini_set ( "session.cookie_lifetime", "7200" );
+	 // 	ini_set ( "session.gc_maxlifetime", "7200" );
+	 // 	session_start ();
+	 // 	$pelicula = R::dispense("usuario");
+		// $pelicula -> admin = "SI";
+		
+		// R::store($pelicula);
+		
+		
+	 // 	R::close();
+	 // 	$this->template->load("plantillaAdmin","administrador/principal", $datos);
+	 // 	//$this->template->load("plantilla","administrador/principal");
+	 // 	// $this->load->view("prueba/Prue",$datos);
+	 // 	//$this->load->view("entrada/aConfirmar");
+	 // }
 	
 	
 }
